requirements-app.txt
# Requirements for the auto Google docs app
panel
PyDrive2
google-auth-httplib2
google-api-python-client
pyprojroot
python-dotenv
fastapi
uvicorn
python-multipart
